+++ src/recorder.ts
@@ -1,6 +1,10 @@
+import { throwError } from './exception/exception';
+import { downloadPCM, downloadWAV } from './download/download';
+import { compress, encodePCM, encodeWAV } from './transform/transform';
+import Player from './player/player';
+
 declare let window: any;
 declare let Math: any;
-declare let document: any;
 declare let navigator: any;
 declare let Promise: any;
 
@@ -10,14 +14,6 @@ interface recorderConfig {
     sampleRate?: number,        // 采样率
     numChannels?: number,       // 声道数
     compiling?: boolean,        // 是否边录边播
-}
-
-interface dataview {
-    byteLength: number,
-    buffer: {
-        byteLength: number,
-    },
-    getUint8: any,
 }
 
 class Recorder {
@@ -41,8 +37,6 @@ class Recorder {
     private analyser: any;
     private littleEdian: boolean;               // 是否是小端字节序
     private prevDomainData: any;                // 存放前一次图形化的数据
-    private playStamp: number = 0;              // 播放录音时 AudioContext 记录的时间戳
-    private playTime: number = 0;               // 记录录音播放时长
     private offset: number = 0;                 // 边录边转，记录外部的获取偏移位置
     private stream: any;                        // 流
 
@@ -57,6 +51,11 @@ class Recorder {
         vol: number,
         data: Array<DataView>,      // 当前存储的所有录音数据
     }) => void;
+    public onplay: () => void;                  // 音频播放回调
+    public onpauseplay: () => void;             // 音频暂停回调
+    public onresumeplay: () => void;            // 音频恢复播放回调
+    public onstopplay: () => void;              // 音频停止播放回调
+    public onplayend: () => void;               // 音频正常播放结束
     /**
      * @param {Object} options 包含以下三个参数：
      * sampleBits，采样位数，一般8,16，默认16
@@ -252,9 +251,10 @@ class Recorder {
         this.source && this.source.stop();
 
         this.isplaying = true;
-        this.playTime = 0;
 
-        this.playAudioData();
+        this.onplay();
+        Player.addPlayEnd(this.onplayend);
+        Player.play(this.getWAV().buffer);
     }
 
     /**
@@ -268,10 +268,9 @@ class Recorder {
             return;
         }
 
-        this.source && this.source.disconnect();
-        // 多次暂停需要累加
-        this.playTime += this.context.currentTime - this.playStamp;
         this.isplaying = false;
+        this.onpauseplay();
+        Player.pausePlay();
     }
 
     /**
@@ -280,13 +279,14 @@ class Recorder {
      * @memberof Recorder
      */
     resumePlay(): void {
-        if (this.isrecording || this.isplaying || 0 === this.playTime) {
+        if (this.isrecording || this.isplaying) {
             // 正在录音或已经播放或没开始播放，恢复无效
             return;
         }
 
         this.isplaying = true;
-        this.playAudioData();
+        this.onresumeplay();
+        Player.resumePlay();
     }
 
     /**
@@ -300,9 +300,9 @@ class Recorder {
             return;
         }
 
-        this.playTime = 0;
         this.isplaying = false;
-        this.source && this.source.stop();
+        this.onstopplay();
+        Player.stopPlay();
     }
 
     /**
@@ -328,30 +328,6 @@ class Recorder {
         this.offset = length;
 
         return data;
-    }
-
-    /**
-     * 利用 decodeAudioData播放录音数据，每次播放都需创建，因为buffersource只能被使用一次
-     *
-     * @private
-     * @memberof Recorder
-     */
-    private playAudioData(): void {
-        this.context.decodeAudioData(this.getWAV().buffer, buffer => {
-            this.source = this.context.createBufferSource();
-
-            // 设置数据
-            this.source.buffer = buffer;
-            // connect到分析器，还是用录音的，因为播放时不能录音的
-            this.source.connect(this.analyser);
-            this.analyser.connect(this.context.destination);
-            this.source.start(0, this.playTime);
-
-            // 记录当前的时间戳，以备暂停时使用
-            this.playStamp = this.context.currentTime;
-        }, function(e) {
-            Recorder.throwError(e);
-        });
     }
 
     /**
@@ -437,9 +413,9 @@ class Recorder {
         // 二维转一维
         let data: any = this.flat();
         // 压缩或扩展
-        data = Recorder.compress(data, this.inputSampleRate, this.outputSampleRate);
+        data = compress(data, this.inputSampleRate, this.outputSampleRate);
         // 按采样位数重新编码
-        return this.PCM = Recorder.encodePCM(data, this.oututSampleBits, this.littleEdian);
+        return this.PCM = encodePCM(data, this.oututSampleBits, this.littleEdian);
     }
 
     /**
@@ -463,7 +439,7 @@ class Recorder {
     downloadPCM(name: string = 'recorder'): void {
         let pcmBlob = this.getPCMBlob();
         
-        this.download(pcmBlob, name, 'pcm');
+        downloadPCM(pcmBlob, name);
     }
 
     /**
@@ -474,7 +450,7 @@ class Recorder {
      */
     private getWAV() {
         let pcmTemp = this.getPCM(),
-            wavTemp = Recorder.encodeWAV(pcmTemp, this.inputSampleRate, 
+            wavTemp = encodeWAV(pcmTemp, this.inputSampleRate, 
                 this.outputSampleRate, this.config.numChannels, this.oututSampleBits, this.littleEdian);
 
         return wavTemp;
@@ -500,8 +476,8 @@ class Recorder {
      */
     downloadWAV(name: string = 'recorder'): void {
         let wavBlob = this.getWAVBlob();
-        
-        this.download(wavBlob, name, 'wav');
+
+        downloadWAV(wavBlob, name);
     }
 
     /**
@@ -515,12 +491,12 @@ class Recorder {
         let lBuffer = new Float32Array(lData),
             rBuffer = new Float32Array(rData);
 
-        let data = Recorder.compress({
+        let data = compress({
             left: lBuffer,
             right: rBuffer,
         }, this.inputSampleRate, this.outputSampleRate);
 
-        return Recorder.encodePCM(data, this.oututSampleBits, this.littleEdian);
+        return encodePCM(data, this.oututSampleBits, this.littleEdian);
     }
 
     /**
@@ -561,26 +537,6 @@ class Recorder {
     }
 
     /**
-     * 下载录音文件
-     * @private
-     * @param {*} blob      blob数据
-     * @param {string} name 下载的文件名
-     * @param {string} type 下载的文件后缀
-     * @memberof Recorder
-     */
-    private download(blob, name: string, type: string): void {
-        try {
-            let oA = document.createElement('a');
-            
-            oA.href = window.URL.createObjectURL(blob);
-            oA.download = name + '.' + type;
-            oA.click();
-        } catch(e) {
-            Recorder.throwError(e);
-        }
-    }
-
-    /**
      * 清空状态，重新开始录音（变量初始化）
      *
      * @private
@@ -596,7 +552,6 @@ class Recorder {
         this.duration = 0;
         this.ispause = false;
         this.isplaying = false;
-        this.playTime = 0;
 
         // 录音前，关闭录音播放
         if (this.source) {
@@ -646,187 +601,6 @@ class Recorder {
             left: lData,
             right: rData
         };
-    }
-
-    /** 
-     * 播放外部音乐文件
-     * 
-     * @param {float32array} blob    blob音频数据
-     * @memberof Recorder
-     */
-    static playAudio(blob): void {
-        let oAudio = document.createElement('audio');
-
-        oAudio.src = window.URL.createObjectURL(blob);
-        // 播放音乐
-        oAudio.play();
-    }
-
-    /**
-     * 数据合并压缩
-     * 根据输入和输出的采样率压缩数据，
-     * 比如输入的采样率是48k的，我们需要的是（输出）的是16k的，由于48k与16k是3倍关系，
-     * 所以输入数据中每隔3取1位
-     * 
-     * @static
-     * @param {float32array} data       [-1, 1]的pcm数据
-     * @param {number} inputSampleRate  输入采样率
-     * @param {number} outputSampleRate 输出采样率
-     * @returns  {float32array}         压缩处理后的二进制数据
-     * @memberof Recorder
-     */
-    static compress(data, inputSampleRate: number, outputSampleRate: number) {
-        // 压缩，根据采样率进行压缩
-        let rate = inputSampleRate / outputSampleRate,
-            compression = Math.max(rate, 1),
-            lData = data.left,
-            rData = data.right,
-            length = Math.floor(( lData.length + rData.length ) / rate),
-            result = new Float32Array(length),
-            index = 0,
-            j = 0;
-
-        // 循环间隔 compression 位取一位数据
-        while (index < length) {
-            let temp = Math.floor(j)
-            
-            result[index] = lData[temp];
-            index++;
-
-            if (rData.length) {
-                /*
-                 * 双声道处理
-                 * e.inputBuffer.getChannelData(0)得到了左声道4096个样本数据，1是右声道的数据，
-                 * 此处需要组和成LRLRLR这种格式，才能正常播放，所以要处理下
-                 */
-                result[index] = rData[temp];
-                index++;
-            }
-            
-            j += compression;
-        }
-        // 返回压缩后的一维数据
-        return result;
-    }
-
-    /**
-     * 转换到我们需要的对应格式的编码
-     * 
-     * @static
-     * @param {float32array} bytes      pcm二进制数据
-     * @param {number}  sampleBits      采样位数
-     * @param {boolean} littleEdian     是否是小端字节序
-     * @returns {dataview}              pcm二进制数据
-     * @memberof Recorder
-     */
-    static encodePCM(bytes, sampleBits: number, littleEdian: boolean = true) {
-        let offset = 0,
-            dataLength = bytes.length * (sampleBits / 8),
-            buffer = new ArrayBuffer(dataLength),
-            data = new DataView(buffer);
-    
-        // 写入采样数据
-        if (sampleBits === 8) {
-            for (let i = 0; i < bytes.length; i++, offset++) {
-                // 范围[-1, 1]
-                let s = Math.max(-1, Math.min(1, bytes[i]));
-                // 8位采样位划分成2^8=256份，它的范围是0-255; 
-                // 对于8位的话，负数*128，正数*127，然后整体向上平移128(+128)，即可得到[0,255]范围的数据。
-                let val = s < 0 ? s * 128 : s * 127;
-                val = +val + 128;
-                data.setInt8(offset, val);
-            }
-        } else {
-            for (let i = 0; i < bytes.length; i++, offset += 2) {
-                let s = Math.max(-1, Math.min(1, bytes[i]));
-                // 16位的划分的是2^16=65536份，范围是-32768到32767
-                // 因为我们收集的数据范围在[-1,1]，那么你想转换成16位的话，只需要对负数*32768,对正数*32767,即可得到范围在[-32768,32767]的数据。
-                data.setInt16(offset, s < 0 ? s * 0x8000 : s * 0x7FFF, littleEdian);
-            }
-        }
-    
-        return data;
-    }
-
-    /**
-     * 编码wav，一般wav格式是在pcm文件前增加44个字节的文件头，
-     * 所以，此处只需要在pcm数据前增加下就行了。
-     * 
-     * @static
-     * @param {DataView} bytes           pcm二进制数据
-     * @param {number}  inputSampleRate  输入采样率
-     * @param {number}  outputSampleRate 输出采样率
-     * @param {number}  numChannels      声道数
-     * @param {number}  oututSampleBits  输出采样位数
-     * @param {boolean} littleEdian      是否是小端字节序
-     * @returns {DataView}               wav二进制数据
-     * @memberof Recorder
-     */
-    static encodeWAV(bytes: dataview, inputSampleRate: number, outputSampleRate: number, numChannels: number, oututSampleBits: number, littleEdian: boolean = true) {
-        let sampleRate = outputSampleRate > inputSampleRate ? inputSampleRate : outputSampleRate,   // 输出采样率较大时，仍使用输入的值，
-            sampleBits = oututSampleBits,
-            buffer = new ArrayBuffer(44 + bytes.byteLength),
-            data = new DataView(buffer),
-            channelCount = numChannels, // 声道
-            offset = 0;
-    
-        // 资源交换文件标识符
-        writeString(data, offset, 'RIFF'); offset += 4;
-        // 下个地址开始到文件尾总字节数,即文件大小-8
-        data.setUint32(offset, 36 + bytes.byteLength, littleEdian); offset += 4;
-        // WAV文件标志
-        writeString(data, offset, 'WAVE'); offset += 4;
-        // 波形格式标志
-        writeString(data, offset, 'fmt '); offset += 4;
-        // 过滤字节,一般为 0x10 = 16
-        data.setUint32(offset, 16, littleEdian); offset += 4;
-        // 格式类别 (PCM形式采样数据)
-        data.setUint16(offset, 1, littleEdian); offset += 2;
-        // 声道数
-        data.setUint16(offset, channelCount, littleEdian); offset += 2;
-        // 采样率,每秒样本数,表示每个通道的播放速度
-        data.setUint32(offset, sampleRate, littleEdian); offset += 4;
-        // 波形数据传输率 (每秒平均字节数) 声道数 × 采样频率 × 采样位数 / 8
-        data.setUint32(offset, channelCount * sampleRate * (sampleBits / 8), littleEdian); offset += 4;
-        // 快数据调整数 采样一次占用字节数 声道数 × 采样位数 / 8
-        data.setUint16(offset, channelCount * (sampleBits / 8), littleEdian); offset += 2;
-        // 采样位数
-        data.setUint16(offset, sampleBits, littleEdian); offset += 2;
-        // 数据标识符
-        writeString(data, offset, 'data'); offset += 4;
-        // 采样数据总数,即数据总大小-44
-        data.setUint32(offset, bytes.byteLength, littleEdian); offset += 4;
-        
-        // 给wav头增加pcm体
-        for (let i = 0; i < bytes.byteLength;) {
-            data.setUint8(offset, bytes.getUint8(i));
-            offset++;
-            i++;
-        }
-    
-        return data;
-    }
-
-    /**
-     * 异常处理
-     * @static
-     * @param {*} message   错误消息
-     * @memberof Recorder
-     */
-    static throwError(message: string) {
-        throw new Error (message);
-    }
-}
-
-/**
- * 在data中的offset位置开始写入str字符串
- * @param {TypedArrays} data    二进制数据
- * @param {Number}      offset  偏移量
- * @param {String}      str     字符串
- */
-function writeString(data, offset, str): void {
-    for (let i = 0; i < str.length; i++) {
-        data.setUint8(offset + i, str.charCodeAt(i));
     }
 }