+++ changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
++ 增加播放音频时的，开始音频播放(onplay)，暂停音频播放(onpauseplay)，恢复音频播放(onresumeplay)，停止音频播放(onstopplay)，音频播放结束回调(onplayend)。
++ 简单拆分文件。
 
 ## [0.5.3] 2019-12-03
 ### Fixed